+++ CHANGELOG.md
@@ -17,6 +17,7 @@ We go to the next version after each release on
 
 - a11y error in GNOME Shell mobile.
 - Activities button error in GNOME Shell mobile.
+- Marked as urgent signal for window demand attention focus feature.
 - `.set_icon_size` is not a function error (Date Menu).
 - Unwanted padding on disabled power icon.